Base packages cleanup (#19)
* pacstrap: Remove ntfs-3g from base packages

It should be completely replaced by the new built-in driver in the
kernel starting with 5.15.

* pacstrap: Remove systemd-sysvcompat from base packages

systemd plans to remove the SysV compatibility layer completely, so this
package should be removed in the future in Arch Linux. Meanwhile,
it is now a dependency for base, so there is no need to specify it
again.