[unpackfsc] Import unpackfsc module from Calamares-extensions
This does some of the same things that unpackfs does, but
with nicer progress reporting, possibly it is faster.
It does not use rsync, so it might not be a drop-in replacement
for unpackfs.

Instead of having one configuration file with multiple items,
each unpackfsc instance unpacks just one thing.